backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.27
email-validator==2.1.0.post1
Pillow==10.2.0
python-dateutil==2.8.2
Flask-CORS==4.0.0
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
requests==2.31.0
alembic==1.13.1
pytz==2024.1
bcrypt==4.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
six==1.16.0 